backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
apscheduler==3.10.4
httpx==0.27.2
torch>=2.2.0
torchaudio>=2.2.0
# Coqui TTS: transformers<5 for compatibility; codec for audio IO.
transformers>=4.36,<5.0
coqui-tts[codec]>=0.27.0
openai-whisper>=20231117
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.27.0
python-multipart>=0.0.6
